requirements.txt
chardet==3.0.4
cycler==0.10.0
Django==1.11.2
django-crispy-forms==1.6.1
django-csvimport==2.11
django-webpack-loader==0.5.0
matplotlib==2.0.2
numpy==1.13.1
pyevtk==1.0.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
